PT-140 Add base presenter for non navigation view(fragment).
