+++ app/src/main/java/com/nimtego/plectrum/presentation/mvp/presenters/general/InformationPresenter.kt
@@ -5,6 +5,7 @@ import com.nimtego.plectrum.domain.interactor.general.InformationInteractor
 import com.nimtego.plectrum.presentation.manger.ChildItemStorage
 import com.nimtego.plectrum.presentation.mvp.model.information_view.SongDetailsModel
 import com.nimtego.plectrum.presentation.mvp.model.main_tab_model.ChildViewModel
+import com.nimtego.plectrum.presentation.mvp.presenters.base.BaseContentPresenter
 import com.nimtego.plectrum.presentation.mvp.presenters.base.BasePresenter
 import com.nimtego.plectrum.presentation.mvp.view.InformationView
 import com.nimtego.plectrum.presentation.navigation.NavigationHandler
@@ -19,37 +20,28 @@ class InformationPresenter
         private val navigationHandler: NavigationHandler,
         private val interactor: InformationInteractor,
         private val itemStorage: ChildItemStorage
-) : BasePresenter<InformationView>() {
+) : BaseContentPresenter<InformationView>() {
+
+    override lateinit var router: Router
 
     private lateinit var navigationQualifier: String
-    private var router: Router? = null
     private var dataModel: ChildViewModel? = null
 
-    override fun attachView(view: InformationView) {
-        super.attachView(view)
-        if(!isInRestoreState(view)) {
-            prepareViewModel()
-            showModel()
-        }
-    }
 
-    private fun prepareViewModel() {
+    override fun prepareViewModel() {
         this.dataModel ?: run {
             this.itemStorage.getCurrentChildItem()?.let {
                 this.dataModel = it
 
             }
         }
+        showModel()
     }
 
     private fun showModel() {
         dataModel?.let {
             viewState.showViewState(it)
         }
-    }
-
-    fun onBackPressed() {
-        this.router?.exit()
     }
 
     fun setNavigationQualifier(navigationQualifier: String) {
