+++ app/src/main/java/com/nimtego/plectrum/presentation/mvp/presenters/popular/BookTabPresenter.kt
@@ -10,6 +10,7 @@ import com.nimtego.plectrum.presentation.mvp.view.TabContentView
 import com.nimtego.plectrum.presentation.mvp.model.main_tab_model.BaseParentViewModel
 import com.nimtego.plectrum.presentation.mvp.model.main_tab_model.ChildViewModel
 import com.nimtego.plectrum.presentation.mvp.model.main_tab_model.ParentTabModelContainer
+import com.nimtego.plectrum.presentation.mvp.presenters.base.BaseContentPresenter
 import com.nimtego.plectrum.presentation.mvp.presenters.base.BasePresenter
 import com.nimtego.plectrum.presentation.navigation.Screens
 import com.nimtego.plectrum.presentation.ui.widget.adapters.ParentTabAdapter
@@ -19,15 +20,15 @@ import javax.inject.Inject
 
 @InjectViewState
 class BookTabPresenter @Inject constructor(
-        private val router: Router,
+        override var router: Router,
         private val itemStorage: MainItemStorage,
         private val interactor: PopularBookInteractor
-) : BasePresenter<TabContentView>(), ParentTabAdapter.OnItemClickListener {
+) : BaseContentPresenter<TabContentView>(), ParentTabAdapter.OnItemClickListener {
 
     private var tabContentModel: BaseParentViewModel<ChildViewModel>? = null
 
 
-    fun viewIsReady() {
+    override fun prepareViewModel() {
         tabContentModel?.let {
             showModel()
         } ?: run {
@@ -69,9 +70,4 @@ class BookTabPresenter @Inject constructor(
                 Screens.ItemInformationScreen(NavigationQualifiers.TAB_BOOK_NAVIGATION)
         )
     }
-
-    fun onBackPressed() {
-        this.router.exit()
-    }
-
 }
