+++ app/src/main/java/com/nimtego/plectrum/presentation/mvp/presenters/search/SearchContentPresenter.kt
@@ -8,6 +8,7 @@ import com.nimtego.plectrum.presentation.manger.UserSearchItemStorage
 import com.nimtego.plectrum.presentation.mvp.model.main_tab_model.ChildViewModel
 import com.nimtego.plectrum.presentation.mvp.model.song.Song
 import com.nimtego.plectrum.presentation.mvp.model.song.SongWrapperModel
+import com.nimtego.plectrum.presentation.mvp.presenters.base.BaseContentPresenter
 import com.nimtego.plectrum.presentation.mvp.presenters.base.BasePresenter
 import com.nimtego.plectrum.presentation.mvp.view.SearchContentView
 import com.nimtego.plectrum.presentation.navigation.NavigationHandler
@@ -24,26 +25,20 @@ class SearchContentPresenter @Inject constructor(
         private val searchItemStorage: UserSearchItemStorage,
         private val userChoiceItemStorage: MainItemStorage,
         private val schedulersProvider: SchedulersProvider
-) : BasePresenter<SearchContentView>(), MoreSectionAdapter.OnItemClickListener {
+) : BaseContentPresenter<SearchContentView>(), MoreSectionAdapter.OnItemClickListener {
+
+    override lateinit var router: Router
 
     private lateinit var navigationQualifier: String
-    private var router: Router? = null
     private var dataModel: List<ChildViewModel>? = null
     private var currentSearchText: String? = null
 
     override fun onUserItemClicked(childViewModel: ChildViewModel) {
         this.userChoiceItemStorage.changeCurrentChildItem(childViewModel)
-        this.router?.navigateTo(Screens.SearchItemInformationScreen(navigationQualifier))
+        this.router.navigateTo(Screens.SearchItemInformationScreen(navigationQualifier))
     }
 
-    override fun attachView(view: SearchContentView) {
-        super.attachView(view)
-        if (!isInRestoreState(view)) {
-            prepareViewModel()
-        }
-    }
-
-    private fun prepareViewModel() {
+    override fun prepareViewModel() {
         this.dataModel?.let { showModel() } ?: run { requestSearchData() }
     }
 
@@ -68,10 +63,6 @@ class SearchContentPresenter @Inject constructor(
         dataModel?.let {
             viewState.showViewState(it)
         }
-    }
-
-    fun onBackPressed() {
-        this.router?.exit()
     }
 
     fun setNavigationQualifier(navigationQualifier: String) {
