+++ app/src/main/java/com/nimtego/plectrum/presentation/ui/fragment/popular/MoreSectionFragment.kt
@@ -48,7 +48,6 @@ class MoreSectionFragment : BaseFragment(), MoreSectionView, BackButtonListener 
     override fun onActivityCreated(savedInstanceState: Bundle?) {
         super.onActivityCreated(savedInstanceState)
         initRV()
-        presenter.prepareViewModel()
     }
 
     protected fun initRV() {
